@@ -19,7 +19,6 @@ var ctxOptionsKey ctxKeyType
 // to skip the commenting on Query.
 //
 //	client.T.Query().All(sqlcomment.Skip(ctx))
-//
 func Skip(ctx context.Context) context.Context {
 	c, ok := ctx.Value(ctxOptionsKey).(*ctxOptions)
 	if !ok {
@@ -31,6 +30,7 @@ func Skip(ctx context.Context) context.Context {
 
 // WithTag stores the key and val pair on the context.
 // for example, if you want to add `route` tag to your SQL comment, put the url path on request context:
+//
 //	middleware := func(next http.Handler) http.Handler {
 //		fn := func(w http.ResponseWriter, r *http.Request) {
 //			ctx := sqlcomment.WithTag(r.Context(), "route", r.URL.Path)
