@@ -1,10 +1,10 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package runtime
 
 // The schema-stitching logic is generated in ariga.io/sqlcomment/examples/ent/runtime.go
 
 const (
-	Version = "v0.9.2-0.20210821141344-368a8f7a2e9a"            // Version of ent codegen.
-	Sum     = "h1:ONZ5BfDxYM5/YMv6cceUXLAkxtCPNSNBUXyVBg6YucU=" // Sum of ent codegen.
+	Version = "v0.13.1"                                         // Version of ent codegen.
+	Sum     = "h1:uD8QwN1h6SNphdCCzmkMN3feSUzNnVvV/WIkHKMbzOE=" // Sum of ent codegen.
 )
